add hook after server has stopped to run memory cleaning fixes
